corrigiendo pequeño error de sintaxis

+++ firmware/src/plantower.cpp
@@ -4,9 +4,10 @@
 #define LENG 31   //0x42 + 23 bytes equal to 24 bytes ->
 unsigned char buf[LENG];
 
-int PM01Value = 0;          //define PM1.0 value of the air detector module
-int PM2_5Value = 0;         //define PM2.5 value of the air detector module
-int PM10Value = 0;         //define PM10 value of the air detector module
+unsigned short int PM01Value = 0;          //define PM1.0 value of the air detector module
+unsigned short int PM2_5Value = 0;         //define PM2.5 value of the air detector module
+unsigned short int PM10Value = 0;         //define PM10 value of the air detector module
+//unsigned short int es en este caso para que sea numeros naturales(por que no va haber -35pm2.5) incluyendo el 0 y una cantiadad no desorbitada de pm25
 
 SoftwareSerial pmsSerial(5, 6);
